refactor(factory): move strategies to the same directory
